CI-FIX: only do special postgres cleanup if the namespace exists
[#175746818](https://www.pivotaltracker.com/story/show/175746818)

+++ ci/tasks/delete-cf/task.sh
@@ -24,10 +24,12 @@ fi
 
 gcloud_auth "${cluster_name}"
 
-echo "Doing some special deletion of postgres resources..."
-set +x
-kubectl delete statefulset cf-db-postgresql -n cf-db
-kubectl delete pod cf-db-postgresql-0 -n cf-db --force --grace-period 0
-set -x
+if kubectl get namespace cf-db &>/dev/null ; then
+  echo "Doing some special deletion of postgres resources..."
+  set +x
+  kubectl delete statefulset cf-db-postgresql -n cf-db
+  kubectl delete pod cf-db-postgresql-0 -n cf-db --force --grace-period 0
+  set -x
+fi
 
 kapp delete -a cf --yes
